Inline the router registration helpers and hold the router client on the cell
